Fix Actor lod levels and material indexes uint32->size_t

+++ Gems/EMotionFX/Code/EMotionFX/Exporters/ExporterLib/Exporter/NodeExport.cpp
@@ -430,7 +430,7 @@ namespace ExporterLib
         MCore::LogInfo("============================================================");
 
         // get all nodes that are affected by the skin
-        AZStd::vector<uint32> bones;
+        AZStd::vector<size_t> bones;
         if (actor)
         {
             actor->ExtractBoneList(0, &bones);
