+++ Gems/EMotionFX/Code/EMotionFX/Source/MeshDeformer.h
@@ -49,7 +49,7 @@ namespace EMotionFX
          * @param node The node where the mesh belongs to during this initialization.
          * @param lodLevel The LOD level of the mesh the mesh deformer works on.
          */
-        virtual void Reinitialize(Actor* actor, Node* node, uint32 lodLevel);
+        virtual void Reinitialize(Actor* actor, Node* node, size_t lodLevel);
 
         /**
          * Creates an exact clone (copy) of this deformer, and returns a pointer to it.
